+++ src/vue/mixins/identity-getter.js
@@ -1,25 +1,36 @@
 import { api } from '@/api'
 import { errors } from '@/js/errors'
+import { mapActions, mapGetters } from 'vuex'
+import { vuexTypes } from '@/vuex'
+
 import AccountGetterMixin from './account-getter'
 
 export default {
   mixins: [AccountGetterMixin],
 
+  computed: {
+    ...mapGetters([
+      vuexTypes.emailByAccountId,
+    ]),
+  },
   methods: {
+    ...mapActions([
+      vuexTypes.LOAD_IDENTITIES_BY_ACCOUNT_ID,
+    ]),
     /**
      * Fetches an account id by email
      *
-     * @param {string} email
+     * @param {string} identifier
      * @param {*} [defaultValue] Value returned if no email found. If not set an
      * instance of `UserDoesntExistError` thrown
      */
-    async getAccountIdByEmail (email, defaultValue) {
-      if (typeof email !== 'string') {
-        throw new TypeError(`getAccountIdByEmail(): 'email' arg should be a string, got ${email}`)
+    async getAccountIdByIdentifier (identifier, defaultValue) {
+      if (typeof identifier !== 'string') {
+        throw new TypeError(`getAccountIdByIdentifier(): 'email' arg should be a string, got ${identifier}`)
       }
 
       const { data } = await api.get('/identities', {
-        filter: { email: email.toLowerCase() },
+        filter: { identifier: identifier.toLowerCase() },
         page: { limit: 1 },
       })
 
@@ -44,7 +55,7 @@ export default {
       // so the user can know which emails were not found
 
       const accountIds = await Promise.all(
-        emails.map(email => this.getAccountIdByEmail(email, null))
+        emails.map(email => this.getAccountIdByIdentifier(email, null))
       )
 
       return accountIds.filter(item => item)
@@ -60,7 +71,7 @@ export default {
      * `UserDoesntExistError` thrown
      */
     async getBalanceIdByEmail (email, assetCode, defaultValue) {
-      const accId = await this.getAccountIdByEmail(email, defaultValue)
+      const accId = await this.getAccountIdByIdentifier(email, defaultValue)
       const balId = await this.getBalanceId(accId, assetCode, defaultValue)
       return balId
     },
@@ -80,15 +91,11 @@ export default {
     },
 
     async getEmailByAccountId (accountId) {
-      const { data } = await api.get('/identities', {
-        filter: { address: accountId },
-        page: { limit: 1 },
-      })
-
-      if (data && data[0]) {
-        return data[0].email
+      if (this.emailByAccountId(accountId)) {
+        return this.emailByAccountId(accountId)
       } else {
-        throw new errors.UserDoesntExistError()
+        await this.LOAD_IDENTITIES_BY_ACCOUNT_ID(accountId)
+        return this.emailByAccountId(accountId)
       }
     },
   },